Add intent integration to expose intent handle API.
